+++ models/Students.js
@@ -64,6 +64,7 @@ router.get('/:id',function(req,res){
 })
 
 router.post('/',upload.single('Student_Photo') , function(req,res){
+    if(req.file){
     if(req.body.id){
     var path = "/" + req.file.path.split("\\").join("/")
     console.log(req.file);
@@ -84,13 +85,22 @@ router.post('/',upload.single('Student_Photo') , function(req,res){
         res.send(student)
     })
 }
+}else{
+    Student.create({
+        id: req.body.id,
+        name: req.body.name,
+        year: req.body.year,
+    }, function(err, student){
+        res.send(student)
+    })
+}
 })
 
 router.delete('/:id', function(req, res){
     Student.findOne({id: req.params.id}, function(err, student){
         if(err){ handleError(err)}
-        else if(!student.Student_Photo) {
-            if(typeof student.Student_Photo !== 'undefined' && student.Student_Photo !== ''){
+        else if(student.Student_Photo) {
+            if(typeof student.Student_Photo !== 'undefined' && student.Student_Photo !== '' && student.Student_Photo.indexOf('/uploads/students/default_student') === -1){
                 fs.unlinkSync('.'+student.Student_Photo);
             }
         }
